portal-system/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.32
alembic==1.13.2
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0,<4.1.0
pydantic-settings==2.4.0
python-multipart==0.0.9
httpx==0.27.0
cryptography==43.0.0
cos-python-sdk-v5>=1.9.0
